SkinAnalysis/requirements.txt
Flask==3.0.1
gunicorn==21.2.0
keras==3.1.1
numpy==1.26.2
tensorflow==2.16.1
pillow==8.1.2